+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+• fix: preserve widget states
+• BREAKING: remove fade_in_out_animation.dart and fade_through_animation.dart
+
 ## 0.1.0
 
 Initial Version of the library.